--- atlasdb-client/src/main/java/com/palantir/atlasdb/cache/EntryMapper.java
@@ -1,26 +1,0 @@
-/*
- * (c) Copyright 2020 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.atlasdb.cache;
-
-import okio.ByteString;
-
-public interface EntryMapper<K, V> {
-    ByteString serializeKey(K key);
-    K deserializeKey(ByteString key);
-    ByteString serializeValue(K key, V value);
-    V deserializeValue(ByteString key, ByteString value);
-}
