+++ atlasdb-client/src/test/java/com/palantir/atlasdb/cache/DeltaEncodingTimestampEntryMapperTests.java
@@ -28,6 +28,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
+import com.palantir.atlasdb.cache.DefaultOffHeapCache.EntryMapper;
 import com.palantir.atlasdb.table.description.ValueType;
 
 import okio.ByteString;
